Fix #4370
Don't use custom metaclass when documenting to work around sphinx documenting __call__ rather than __init__
