+++ docs/conf.py
@@ -18,6 +18,8 @@
 #
 import os
 import sys
+from abc import ABCMeta
+from importlib import reload
 
 # Import matplotlib and set the backend
 # before qcodes imports pyplot and automatically
@@ -25,6 +27,19 @@ import sys
 import matplotlib
 import sphinx_rtd_theme
 from packaging.version import parse
+
+# setting the metaclass will cause sphinx
+# to document the signature of `__call__`
+# rather than `__init__` that is unhelpful
+# for instruments. When building the docs
+# we patch it back to ABCMeta
+import qcodes.instrument.instrument_meta
+
+qcodes.instrument.instrument_meta.InstrumentMeta = ABCMeta
+reload(qcodes.instrument.instrument)
+reload(qcodes.instrument.ip)
+reload(qcodes.instrument.visa)
+reload(qcodes.instrument)
 
 import qcodes
